+++ app/package.json
@@ -32,6 +32,7 @@
     "@talisman-connect/wallets": "^1.1.3",
     "@talismn/connect-components": "^1.1.8",
     "@tanstack/react-query": "^5.36.2",
+    "@types/lodash.debounce": "^4.0.9",
     "@vercel/analytics": "^1.3.1",
     "@web3modal/wagmi": "^4.2.1",
     "class-variance-authority": "^0.7.0",
@@ -39,6 +40,7 @@
     "ethers": "^6.12.1",
     "flatted": "^3.3.1",
     "framer-motion": "^11.1.9",
+    "lodash.debounce": "^4.0.8",
     "lucide-react": "^0.394.0",
     "next": "14.2.3",
     "react": "^18",
